chore(*): update element registration functions
