+++ README.md
@@ -76,7 +76,7 @@ yarn link "@kong/kong-auth-elements"
 2. The first tag within the `<template>` of a custom element `*.ce.vue` file must be `<div class="kong-auth-element">` and it must wrap all other content within the template.
 3. Custom elements must utilize the `<BaseCustomElement />` component as the first child of the `<div class="kong-auth-element">` element which will wrap any other structure/components (this enables style injection for child components).
 4. Custom elements must be added to the following path `/src/elements/{kebab-case-element-name}/{PascalCaseElementName}.ce.vue`
-5. Custom elements must have an `index.ts` file at the root of their `/src/elements/{kebab-case-element-name}/` directory that exports a registration function.
+5. Custom elements must have an export function added in the `/src/elements/index.ts` file that exports the registration function for the element. Note the proper names/paths in the file.
 6. Custom element templates (the contents of the `{PascalCaseElementName}.ce.vue` file) must utilize the template shown below:
 
     <details>
